requirements.txt
Flask
tensorflow
numpy
Pillow
opencv-python
scikit-image
scikit-learn
matplotlib
seaborn
